requirements.txt
catboost==1.2.3
imbalanced-learn==0.12.0
imblearn==0.0
langchain==0.1.13
langchain-openai==0.1.1
munch==4.0.0
numpy==1.20.3
openai==1.14.3
pandas==1.3.4
python-dotenv==1.0.1
scikit-learn==1.3.1
scipy==1.10.0
tabulate==0.9.0
torch==1.13.0
torchvision==0.14.0
tqdm
xgboost==2.0.3
lightgbm